Removed old debug code
There was debug code in the bezier%init method that wrote
the interface control points to a file. This code assumed that
the grid was 2 blocks x 2 blocks. This causes a segmentation fault
when a block dimension is less than 2, and doesn't write all
control points to file when a block dimension is greater than 2.
I removed this code, as it was only there for debugging, anyhow.
I also cleaned up some new lines and removed unnecessary files.